// src/modules/preaching/interfaces/preaching.ts
export interface Preaching {
    id: string;
    userId: string;
    day: string;
    initHour: string;
    finalHour: string;
    createdAt: string;
    updatedAt: string;
}
// ...
export interface PreachingEndpoint {
    id: string;
    user_id: string;
    day: string;
    init_hour: string;
    final_hour: string;
    created_at: string;
    updated_at: string;
}

/**
 * Defining the structure of the PreachingFormValues object.
 *
 * @property {Date} day - This is the day of the preaching.
 * @property {Date} initHour - This is the initial hour of the preaching.
 * @property {Date} finalHour - This is the final hour of the preaching.
 */
export interface PreachingFormValues {
    day: Date;
    initHour: Date;
    finalHour: Date;
}